+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Added
+- Added support for the `aarch64-pc-windows-msvc` target.
+
 ## [0.8.6] - 2022-11-27
 ### Added
 - Added binaries for Windows targets using the GNU toolchain.